# docs/deliverables/system_docs_20260729/star-rail-agents.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Star Rail Agents",
    "description": "基于 Multi-Agent 与 RAG 的游戏智能助手平台",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/profile/characters": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "Get Character Pool",
        "operationId": "get_character_pool_api_v1_profile_characters_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterPoolResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "profile"
        ],
        "summary": "Replace Character Pool",
        "operationId": "replace_character_pool_api_v1_profile_characters_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterPoolUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterPoolResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/profile/characters/{character_id}/favorite": {
      "patch": {
        "tags": [
          "profile"
        ],
        "summary": "Update Favorite Character",
        "operationId": "update_favorite_character_api_v1_profile_characters__character_id__favorite_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "character_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Character Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FavoriteCharacterUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCharacterResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/profile/characters/{character_id}/progression": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "Get Character Progress",
        "operationId": "get_character_progress_api_v1_profile_characters__character_id__progression_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "character_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Character Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterProgressResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "profile"
        ],
        "summary": "Save Character Progress",
        "operationId": "save_character_progress_api_v1_profile_characters__character_id__progression_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "character_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Character Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterProgressUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterProgressResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/stories": {
      "get": {
        "tags": [
          "stories"
        ],
        "summary": "List Stories",
        "operationId": "list_stories_api_v1_stories_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 100,
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "version",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 16,
              "default": "",
              "title": "Version"
            }
          },
          {
            "name": "world",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 80,
              "default": "",
              "title": "World"
            }
          },
          {
            "name": "mission_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 40,
              "default": "",
              "title": "Mission Type"
            }
          },
          {
            "name": "series",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 120,
              "default": "",
              "title": "Series"
            }
          },
          {
            "name": "character",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 48,
              "default": "",
              "title": "Character"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stories/{mission_id}": {
      "get": {
        "tags": [
          "stories"
        ],
        "summary": "Get Story",
        "operationId": "get_story_api_v1_stories__mission_id__get",
        "parameters": [
          {
            "name": "mission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Mission Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stories/{mission_id}/scenes/{chunk_id}": {
      "get": {
        "tags": [
          "stories"
        ],
        "summary": "Get Story Scene",
        "operationId": "get_story_scene_api_v1_stories__mission_id__scenes__chunk_id__get",
        "parameters": [
          {
            "name": "mission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Mission Id"
            }
          },
          {
            "name": "chunk_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chunk Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryScene"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CharacterPoolResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/UserCharacterResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "favorites": {
            "type": "integer",
            "title": "Favorites"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "favorites"
        ],
        "title": "CharacterPoolResponse"
      },
      "CharacterPoolUpdate": {
        "properties": {
          "character_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 200,
            "title": "Character Ids"
          }
        },
        "type": "object",
        "title": "CharacterPoolUpdate"
      },
      "CharacterProgressResponse": {
        "properties": {
          "current_level": {
            "type": "integer",
            "maximum": 80.0,
            "minimum": 1.0,
            "title": "Current Level"
          },
          "target_level": {
            "type": "integer",
            "maximum": 80.0,
            "minimum": 1.0,
            "title": "Target Level"
          },
          "eidolon": {
            "type": "integer",
            "maximum": 6.0,
            "minimum": 0.0,
            "title": "Eidolon",
            "default": 0
          },
          "current_skills": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Current Skills"
          },
          "target_skills": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Target Skills"
          },
          "character_id": {
            "type": "string",
            "title": "Character Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "current_level",
          "target_level",
          "character_id",
          "created_at",
          "updated_at"
        ],
        "title": "CharacterProgressResponse"
      },
      "CharacterProgressUpdate": {
        "properties": {
          "current_level": {
            "type": "integer",
            "maximum": 80.0,
            "minimum": 1.0,
            "title": "Current Level"
          },
          "target_level": {
            "type": "integer",
            "maximum": 80.0,
            "minimum": 1.0,
            "title": "Target Level"
          },
          "eidolon": {
            "type": "integer",
            "maximum": 6.0,
            "minimum": 0.0,
            "title": "Eidolon",
            "default": 0
          },
          "current_skills": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Current Skills"
          },
          "target_skills": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Target Skills"
          }
        },
        "type": "object",
        "required": [
          "current_level",
          "target_level"
        ],
        "title": "CharacterProgressUpdate"
      },
...
      "FavoriteCharacterUpdate": {
        "properties": {
          "is_favorite": {
            "type": "boolean",
            "title": "Is Favorite"
          }
        },
        "type": "object",
        "required": [
          "is_favorite"
        ],
        "title": "FavoriteCharacterUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StoryDetail": {
        "properties": {
          "mission_id": {
            "type": "string",
            "title": "Mission Id"
          },
          "mission_name": {
            "type": "string",
            "title": "Mission Name"
          },
          "mission_type": {
            "type": "string",
            "title": "Mission Type"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "world": {
            "type": "string",
            "title": "World"
          },
          "series_name": {
            "type": "string",
            "title": "Series Name"
          },
          "mission_order": {
            "type": "integer",
            "title": "Mission Order"
          },
          "characters": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Characters"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "scene_count": {
            "type": "integer",
            "title": "Scene Count",
            "default": 0
          },
          "matched_character": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matched Character"
          },
          "matched_scene_count": {
            "type": "integer",
            "title": "Matched Scene Count",
            "default": 0
          },
          "first_matching_chunk_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Matching Chunk Id"
          },
          "story_text": {
            "type": "string",
            "title": "Story Text"
          },
          "source_url": {
            "type": "string",
            "title": "Source Url"
          },
          "previous_mission": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Mission"
          },
          "next_mission": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Mission"
          },
          "scenes": {
            "items": {
              "$ref": "#/components/schemas/StorySceneSummary"
            },
            "type": "array",
            "title": "Scenes"
          }
        },
        "type": "object",
        "required": [
          "mission_id",
          "mission_name",
          "mission_type",
          "version",
          "world",
          "series_name",
          "mission_order",
          "summary",
          "story_text",
          "source_url"
        ],
        "title": "StoryDetail"
      },
      "StoryFilters": {
        "properties": {
          "versions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Versions"
          },
          "worlds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Worlds"
          },
          "mission_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Mission Types"
          },
          "series": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Series"
          }
        },
        "type": "object",
        "title": "StoryFilters"
      },
      "StoryListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/StorySummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "filters": {
            "$ref": "#/components/schemas/StoryFilters"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "filters"
        ],
        "title": "StoryListResponse"
      },
      "StoryScene": {
        "properties": {
          "chunk_id": {
            "type": "string",
            "title": "Chunk Id"
          },
          "chunk_order": {
            "type": "integer",
            "title": "Chunk Order"
          },
          "scene_title": {
            "type": "string",
            "title": "Scene Title"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "characters": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Characters"
          },
          "mission_id": {
            "type": "string",
            "title": "Mission Id"
          },
          "mission_name": {
            "type": "string",
            "title": "Mission Name"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "source_url": {
            "type": "string",
            "title": "Source Url"
          },
          "previous_chunk_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Chunk Id"
          },
          "next_chunk_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Chunk Id"
          }
        },
        "type": "object",
        "required": [
          "chunk_id",
          "chunk_order",
          "scene_title",
          "mission_id",
          "mission_name",
          "content",
          "source_url"
        ],
        "title": "StoryScene"
      },
...
      "UserCharacterResponse": {
        "properties": {
          "character_id": {
            "type": "string",
            "title": "Character Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "rarity": {
            "type": "integer",
            "title": "Rarity"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "element": {
            "type": "string",
            "title": "Element"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "level": {
            "type": "integer",
            "title": "Level"
          },
          "eidolon": {
            "type": "integer",
            "title": "Eidolon"
          },
          "is_favorite": {
            "type": "boolean",
            "title": "Is Favorite"
          },
          "is_built": {
            "type": "boolean",
            "title": "Is Built"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "character_id",
          "name",
          "rarity",
          "path",
          "element",
          "level",
          "eidolon",
          "is_favorite",
          "is_built",
          "created_at",
          "updated_at"
        ],
        "title": "UserCharacterResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "本地 FastAPI"
    }
  ]
}